Refactored too much code Images are now stored as PNGs in private storage so they should be always available once set initially
