+++ app/src/main/res/layout/activity_image_zone_selection.xml
@@ -2,6 +2,6 @@
                 xmlns:tools="http://schemas.android.com/tools"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                tools:context="com.jamessimshaw.wallpaperhelper.Activities.ImageZoneSelectionActivity">
+                tools:context="com.jamessimshaw.wallpaperhelper.activities.ImageZoneSelectionActivity">
 
 </RelativeLayout>
