+++ app/src/main/res/menu/menu_image_zone_selection.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="com.jamessimshaw.wallpaperhelper.Activities.ImageZoneSelectionActivity">
+      tools:context="com.jamessimshaw.wallpaperhelper.activities.ImageZoneSelectionActivity">
     <item android:id="@+id/action_settings"
           android:title="@string/action_settings"
           android:orderInCategory="100"
